+++ tests/base.py
@@ -12,6 +12,7 @@ from importlib import import_module
 from inspect import getfile
 from shutil import rmtree, copy
 import time
+from unittest import skip, skipIf
 
 import django
 from django import test
@@ -22,10 +23,10 @@ from django.db import models, connection
 from django.db.models.fields import related
 from django.contrib.contenttypes.models import ContentType
 from django.utils.six import StringIO
+from django.apps.registry import apps
 
 from gm2m import GM2MField
 
-from .compat import apps, skip, skipIf
 from .helpers import app_mod_path, del_app_models
 
 # patches the migration questioner
@@ -177,15 +178,7 @@ class TestCase(_TestCase):
         __, __, args, kwargs = field.deconstruct()
         new_field = GM2MField(*args, **kwargs)
 
-        for attr in (''):
-            self.assertEqual(getattr(field, attr),
-                             getattr(new_field, attr))
-
-        for attr in (''):
-            self.assertEqual(getattr(field.rel, attr),
-                             getattr(new_field.rel, attr))
-
-        # just checking the stings output, as for an attr to attr comparison
+        # just checking the strings output, as for an attr to attr comparison
         # we would need to run contribute_to_class
         self.assertSetEqual(set(['%s.%s' % (r.to._meta.app_label,
                                             r.to._meta.object_name)