+++ tests/through/models.py
@@ -1,18 +1,24 @@
 from django.db import models
 from django.contrib.contenttypes.models import ContentType
+from django.contrib.contenttypes.fields import GenericForeignKey
 
 import gm2m
-from gm2m.compat import GenericForeignKey
 
 from ..app.models import Project
 
 
 class Links(models.Model):
 
+    class Meta:
+        app_label = 'through'
+
     related_objects = gm2m.GM2MField(Project, through='RelLinks')
 
 
 class RelLinks(models.Model):
+
+    class Meta:
+        app_label = 'through'
 
     links = models.ForeignKey(Links)
     target = GenericForeignKey(ct_field='target_ct', fk_field='target_fk')